Fetching Posts from API
Fetching the posts from Jsonplaceholder API. Cleaning out a few niggles with the action dispatching the entire response instead of the relevant data. Using switch case instead of if condition in the reducer. Adding the mapStateToProps in the PostList component. Removing our dummy Reducer to wire up the actual reducer in the combineReducers method.